lost_found_reunion2/requirements.txt
streamlit
torch
open_clip_torch
chromadb
pillow
numpy
pandas
